ReadOverlaysFile: Rewrite for efficient overlay state checks
This commit deprecates the old way of ReadOverlaysFile, which was to
duplicate the overlays.xml file in /data/system and parses the XML
language. This was proven to not be as useful as the overlays.xml file
could report back incorrect information depending on the user.

“om list” is an efficient and reliable way to deduce what overlays are
enabled, disabled or non-approved without the use of root (to access
/data/system) as well as without the need of copying files.

This will cut down the process to a fraction of the previous time, as
well as alleviate the I/O and root calls on the device.