InformationActivity: Introduce a less memory intensive hero load
When launching InformationActivity, one of the most time consuming
parts is to extract the bitmap out of the APK and then loading the
image, all before the user sees the activity.

This should not be the case and thus this commit allows for the images
to be loaded asynchronously through a byte[] after a conversion from a
bitmap drawable. This will also allow the InformationActivity to speed
up launching by 2x.