SubstratumUI: Revive functionality to refresh overlays on demand
When a user uninstalls a theme pack, the home fragment (and all
subsequent fragments) should automatically refresh, thus cleaning up
the workspace.

Prior to this commit, Substratum has been using its failsafe of
checking whether the package exists when the user clicks on a card.
This will fall back as the failsafe again as this new implementation is
fully automated.